updated pagination for hangouts, facts
